agentic_search_api/pyproject.toml
[project]
name = "sse"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dotenv>=0.9.9",
    "fastapi>=0.116.1",
    "openai-agents>=0.2.4",
    "uvicorn>=0.35.0",
]
